bt_bert_model/requirements.txt
transformers>=4.40
torch>=2.2
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
pyyaml>=6.0
tqdm>=4.65
